% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_abundance.R
\name{AA_abundance}
\alias{AA_abundance}
\title{Count amino acid abundances in protein sequence and yield barplot :)}
\usage{
AA_abundance(prot_seq)
}
\arguments{
\item{prot_seq}{Protein sequence for abundance count}
}
\value{
Barplot of AA abundances
}
\description{
Count amino acid abundances in protein sequence and yield barplot :)
}
