% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Codon Table}
\format{
Vector of 64 codons with the corresponding amino acid
}
\source{
\url{https://learn.inside.dtu.dk/content/enforced/102974-OFFERING-653081/lab08/lab_08_r_package_exercises.html}
}
\usage{
codon_table
}
\description{
Table of DNA codons to amino acids
}
\keyword{datasets}
