% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_generator.R
\name{DNA_generator}
\alias{DNA_generator}
\title{DNA sequence generator}
\usage{
DNA_generator(DNA_length)
}
\arguments{
\item{DNA_length}{Length of generated DNA}
}
\value{
A generated DNA sequence
}
\description{
DNA sequence generator
}
\examples{
DNA_generator(36)
}
