% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_codon.R
\name{export_amino_acid}
\alias{export_amino_acid}
\title{Title
export amino_acid}
\usage{
export_amino_acid(codons)
}
\arguments{
\item{codons}{}
}
\value{
amino_acid
}
\description{
Title
export amino_acid
}
\examples{
paste0(codon_table['AGU'], collapse = "")
"S"

}
