% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_to_codons.R
\name{sequence_to_codons}
\alias{sequence_to_codons}
\title{Convert sequence to codons}
\usage{
sequence_to_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{RNA sequence}

\item{start}{The reading frame index}
}
\value{
Vector of codons
}
\description{
Convert sequence to codons
}
\examples{
sequence_to_codons("UUGUCUAUCGAA", start = 1)
# [1] "UUG" "UCU" "AUC" "GAA"
}
