% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Translates DNA to RNA}
\usage{
DNA_to_RNA(DNA_seq)
}
\arguments{
\item{DNA_seq}{String of DNA sequence}
}
\value{
A RNA sequence
}
\description{
Translates DNA to RNA
}
\examples{
# "ATGC" > "AUGC"
}
